chore(DataTable): remove tonumber, restructure files and pass columns data as well
