+++ packages/DataTable/package.json
@@ -31,7 +31,6 @@
     "lodash.isequal": "^4.5.0",
     "lodash.ismatchwith": "^4.4.0",
     "lodash.isstring": "^4.0.1",
-    "lodash.tonumber": "^4.0.3",
     "nanoid": "^2.1.7",
     "prop-types": "^15.7.2",
     "react-window": "^1.8.5"
