+++ packages/DataTable/src/context.js
@@ -3,9 +3,7 @@ import PropTypes from "prop-types";
 import isMatchWith from "lodash.ismatchwith";
 import tableReducer from "./reducers/table";
 import useAsyncReducer from "./hooks/useAsyncReducer";
-import filterTesters from "./helpers/filterTesters";
-import sort from "./helpers/sort";
-import { columnTypes, logicalFilterOperators } from "./constants";
+import { logicalFilterOperators } from "./constants";
 
 const TableStateContext = React.createContext();
 const TableDispatchContext = React.createContext();
@@ -92,71 +90,6 @@ function useDispatch() {
   return React.useContext(TableDispatchContext);
 }
 
-function useData() {
-  const {
-    isControlled,
-    data,
-    columns,
-    filters,
-    sortColumn,
-    sortDirection,
-    keygen,
-    logicalFilterOperator,
-  } = useDataTableState();
-  let sortedData = [];
-  let filteredData = [];
-
-  function calculateResult() {
-    if (isControlled) return data;
-    if (sortColumn && sortDirection) {
-      sortedData = sort({
-        data,
-        columnId: sortColumn,
-        direction: sortDirection,
-        columnType:
-          columns[sortColumn].type || (typeof data[0][sortColumn] === "number" ? columnTypes.NUMBER : columnTypes.TEXT),
-        momentParsingFormat: columns[sortColumn].momentParsingFormat,
-      });
-    }
-
-    if (filters.length > 0) {
-      filteredData = data.filter(row => {
-        // checking if filter.rule exist, will removed after having all the rules
-        const tester = filter =>
-          filter.rule ? filterTesters[filter.rule](row[filter.columnId], filter.value, columns[filter.columnId]) : true;
-
-        switch (logicalFilterOperator) {
-          case logicalFilterOperators.AND:
-            return filters.every(tester);
-          case logicalFilterOperators.OR:
-            return filters.some(tester);
-          default:
-            return true;
-        }
-      });
-    }
-
-    if (sortedData.length === 0 && filters.length === 0) return data;
-
-    if (sortedData.length > 0 && filteredData.length > 0)
-      return sortedData.filter(row => filteredData.find(filteredRow => row[keygen] === filteredRow[keygen]));
-
-    return sortedData.length > 0 ? sortedData : filteredData;
-  }
-
-  const result = React.useMemo(calculateResult, [
-    data,
-    columns,
-    filters,
-    sortColumn,
-    sortDirection,
-    keygen,
-    logicalFilterOperator,
-  ]);
-
-  return result;
-}
-
 TableProvider.propTypes = {
   children: PropTypes.node.isRequired,
   isControlled: PropTypes.bool.isRequired,
@@ -175,4 +108,4 @@ TableProvider.defaultProps = {
   columns: [],
 };
 
-export { TableProvider, useDataTableState, useDispatch, useData };
+export { TableProvider, useDataTableState, useDispatch };
